Update: Added Avatar system; Added basic content to My Account
- Added Active indicator component; implemented in the accounts table
- Added Avatar model, basic support systems to the enraiged namespace
  - Ability to select avatar css color
  - Ability to upload,remove image
- Added Avatar,Status ui components; implemented in the accounts table
- Added AvatarEditor component to manage an avatar
- Added Auth store for AppState auth.user
- Added basic content to the My Account (Show) dashboard
- Added forms,pages,routes,etc for updating My {Avatar,Login,Profile}
- Added list view to the My Files display
- Corrected UpdateAccount service issue when password is null
- Corrected issue with table exporting symbol not falsified
- Corrected menu toggle transition (not applied) bug
- Corrected table issues preserving,restoring column sort

+++ packages/enraiged/src/Files/Models/File.php
@@ -5,6 +5,8 @@ namespace Enraiged\Files\Models;
 use Enraiged\Support\Traits\CreatedBy;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\File as IlluminateFile;
+use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 //use Symfony\Component\HttpFoundation\File\File as SymfonyFile;
@@ -58,9 +60,11 @@ class File extends Model
      */
     public function delete()
     {
-        Storage::delete($this->path);
+        if ($this->path) {
+            Storage::delete($this->path);
 
-        return parent::delete();
+            return parent::delete();
+        }
     }
 
     /**
@@ -83,5 +87,41 @@ class File extends Model
     public function inline(): StreamedResponse
     {
         return Storage::response($this->path);
+    }
+
+    /**
+     *  
+     *  @param  UploadedFile $file
+     *  @return self
+     *  @throws type
+     */
+    public function upload(UploadedFile $file): self
+    {
+        DB::transaction(function () use ($file) {
+            $folder = $this->attachable->folder;
+
+            if (!Storage::has($folder)) {
+                Storage::makeDirectory($folder);
+            }
+
+            $extension = $file->guessExtension();
+            $hashname = uhash();
+            $filename = "{$hashname}.{$extension}";
+
+            $parameters = [
+                'name' => $file->getClientOriginalName(),
+                'path' => "{$folder}/{$filename}",
+                'size' => $file->getSize(),
+                'mime' => $file->getMimeType(),
+            ];
+
+            (object) $this
+                ->fill($parameters)
+                ->save();
+
+            $file->storeAs($folder, $filename);
+        });
+
+        return $this;
     }
 }
